payment functionality, formatting

+++ src/schema/money/tag.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Tag {
-    id: u64,
+    id: usize,
     pub content: String,
 }
